rename: 폴더명 변경 및 목업 데이터 삭제
폴더명 변경 및 목업 데이터 삭제
